// src/gameEngine/types.ts
export enum GameStates {
  GAME_STARTING = 'GAME_STARTING',
  GAME_IN_PROGRESS = 'GAME_IN_PROGRESS',
  GAME_STOPPED = 'GAME_STOPPED',
  GAME_ENDED = 'GAME_ENDED',
}

export enum GameEvents {
  GAME_STATE_CHANGED = 'GAME_STATE_CHANGED',
  BANKROLL_CHANGED = 'BANKROLL_CHANGED',
  BET_STATUS_CHANGED = 'BET_STATUS_CHANGED',
  PLAYERS_CHANGED = 'PLAYERS_CHANGED',
  HISTORY_CHANGED = 'HISTORY_CHANGED',
}